refactor(components): move types
