+++ src/assets/billing.ts
@@ -1,0 +1,6 @@
+const BILLING = {
+  monthly: 'mo',
+  yearly: 'yr',
+};
+
+export default BILLING;
