+++ src/types/addon.ts
@@ -1,0 +1,8 @@
+import { PriceByBilling } from './billing';
+
+export type AddonType = {
+  id: string;
+  title: string;
+  description: string;
+  price: PriceByBilling;
+};
